feat: 선택 훅 연결 및 타입 정리
