+++ ㅁ.md
@@ -1,0 +1,3 @@
+[![trophy](https://github-profile-trophy.vercel.app/?username=kojesung&theme=onedark)](https://github.com/ryo-ma/github-profile-trophy)
+![Achievements](https://github-profile-achievements.vercel.app/api/v1/github-profile-achievements?username=kojesung&theme=dark)![GitHub Profile Rank](https://github-readme-stats.vercel.app/api?username=kojesung&show_icons=true&rank_icon=github)
+![GitHub Contributor](https://img.shields.io/badge/GitHub-Contributor-green?style=flat-square&logo=github)
